+++ vehicle-python/src/vehicle_lang/loss/_abc/_translation.py
@@ -62,6 +62,8 @@ class ABCTranslation(
                 return self.translate_MinRatTensor(expression)
             case vcl_ast.MaxRatTensor():
                 return self.translate_MaxRatTensor(expression)
+            case vcl_ast.PowRatTensor():
+                return self.translate_PowRatTensor(expression)
             case vcl_ast.NegRatTensor():
                 return self.translate_NegRatTensor(expression)
             case vcl_ast.ReduceAddRatTensor():
@@ -136,6 +138,11 @@ class ABCTranslation(
     @abstractmethod
     def translate_MaxRatTensor(
         self, expression: vcl_ast.MaxRatTensor
+    ) -> vcl_var.Expression: ...
+
+    @abstractmethod
+    def translate_PowRatTensor(
+        self, expression: vcl_ast.PowRatTensor
     ) -> vcl_var.Expression: ...
 
     @abstractmethod