+++ vehicle/src/Vehicle/Backend/ITP/Agda.hs
@@ -599,7 +599,7 @@ compileBuiltinFunction f args = case f of
   ForeachVector -> annotateApp [VehicleUtils] Nothing "foreachVector" args
   StackTensor {} -> annotateApp [DataTensor] Nothing "stack" args
   Iterate -> unsupportedError
-  PowRat -> unsupportedError
+  Pow {} -> unsupportedError
   where
     unsupportedError :: a
     unsupportedError =