+++ vehicle/src/Vehicle/Data/Builtin/Interface/Blocked.hs
@@ -65,7 +65,7 @@ functionBlockingStatus b spine = case b of
   Div DivRatTensor -> fixedStatus [1, 2] spine
   Min MinRatTensor -> fixedStatus [1, 2] spine
   Max MaxRatTensor -> fixedStatus [1, 2] spine
-  PowRat -> fixedStatus [0, 1] spine
+  Pow PowRatTensor -> fixedStatus [1, 2] spine
   CompareIndex _op -> fixedStatus [2, 3] spine
   CompareNat _op -> fixedStatus [0, 1] spine
   CompareRatTensorPointwise _op -> fixedStatus [1, 2] spine